feat: 마이페이지 (시음 후기 모달) 구현 (#102)
- 응집도 개선
- 스크롤 스타일 수정
