+++ src/App.css
@@ -17,3 +17,10 @@ body {
     'Segoe UI Symbol',
     sans-serif;
 }
+
+@layer utilities {
+  .review-modal-scroll {
+    scrollbar-width: thin;
+    scrollbar-color: rgba(0, 0, 0, 0.25) transparent;
+  }
+}
